:sparkles: api: move certification center repository to  src/certification/shared directory
