+++ api/tests/integration/domain/usecases/create-certification-center-membership-for-sco-organization-member_test.js
@@ -1,7 +1,7 @@
 import { expect, databaseBuilder, knex } from '../../../test-helper.js';
 
 import * as membershipRepository from '../../../../lib/infrastructure/repositories/membership-repository.js';
-import * as certificationCenterRepository from '../../../../lib/infrastructure/repositories/certification-center-repository.js';
+import * as certificationCenterRepository from '../../../../src/certification/shared/infrastructure/repositories/certification-center-repository.js';
 import * as certificationCenterMembershipRepository from '../../../../lib/infrastructure/repositories/certification-center-membership-repository.js';
 
 import { Membership } from '../../../../lib/domain/models/Membership.js';
